+++ CHANGELOG.md
@@ -12,6 +12,8 @@
     - Initial support for `arm64` architecture
     - Allow `$schema` property
 - **CI**:
+    - Files with multiple empty lines at the end now produce error
+    - `UTF8-Bom`, `UTF16 BE`, `UTF16 LE` files are prohibited
     - Support basic validation of yml typed manifests
     - Support validation of all archived manifests
 - **scoop-cat**: Add `-f`, `--format` options